Minor code cleanup

+++ processing/src/main/java/io/druid/query/aggregation/HistogramBufferAggregator.java
@@ -56,7 +56,7 @@ public class HistogramBufferAggregator implements BufferAggregator
   @Override
   public void aggregate(ByteBuffer buf, int position)
   {
-    final float value = (float)selector.get();
+    final float value = selector.get();
     final int minPos = position + minOffset;
     final int maxPos = position + maxOffset;
